Fix QLoRA NaN loss and sorry false positives
QLoRA training (train_deepseek_qlora.py):
- Switch bnb_4bit_compute_dtype from float16 → float32
- V100 doesn't support bfloat16; fp16 compute caused NaN gradients during
  backprop through dequantized 4-bit layers. float32 compute is ~2x slower
  but fully stable. Weights remain in 4-bit (3.5GB), so VRAM still fits 16GB.

SorryDB false positives (tactic_model.py):
- Filter entire proof if any tactic line contains `\bsorry\b` — sorry is
  valid Lean syntax (axiom skip) and passes lake build, but is not a proof
- Filter lines matching LLM meta-prompts ('Complete the following Lean 4 code:')
  which leaked into DeepSeek output and were incorrectly accepted

+++ scripts/train_deepseek_qlora.py
@@ -143,10 +143,12 @@ def main():
 
     # ── Load model in 4-bit NF4 ─────────────────────────────────────────────
     logger.info("Loading model in 4-bit NF4...")
+    # Use float32 compute on V100 — bfloat16 unavailable, fp16 causes NaN gradients
+    # in backprop through 4-bit dequantization. float32 is ~2x slower but stable.
     bnb_config = BitsAndBytesConfig(
         load_in_4bit=True,
         bnb_4bit_quant_type="nf4",
-        bnb_4bit_compute_dtype=torch.float16,
+        bnb_4bit_compute_dtype=torch.float32,
         bnb_4bit_use_double_quant=True,
     )
     model = AutoModelForCausalLM.from_pretrained(
